+++ src/exo/API_cursors.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
 
-from typing import Optional, List, Any
+from typing import List, Any
 
 from . import API  # TODO: remove this circular import
 from .API_types import ExoType, loopir_type_to_exotype
@@ -853,85 +853,6 @@ class StrideExprCursor(ExprCursor):
 
 # --------------------------------------------------------------------------- #
 # --------------------------------------------------------------------------- #
-# High-level cursor navigation functions
-# --------------------------------------------------------------------------- #
-# These could all be user-level code, but are convenient enough to include.
-
-
-class CursorNavigationError(Exception):
-    pass
-
-
-def validate_cursors(c1, c2):
-    """
-    Checks that [c1] and [c2] are both valid cursors, and that they originate from the
-    same proc.
-    """
-    assert not isinstance(c1, InvalidCursor), "first cursor was an InvalidCursor"
-    assert not isinstance(c2, InvalidCursor), "second cursor was an InvalidCursor"
-    assert c1.proc() == c2.proc(), "cursors originate from different procs"
-
-
-def match_level(cursor, cursor_to_match):
-    """
-    Lifts [cursor] through the AST until [cursor] and [cursor_to_match] are at
-    the same level, e.g. have the same parent. Returns an InvalidCursorError if
-    [cursor_to_match]'s parent is not an ancestor of [cursor].
-    """
-    validate_cursors(cursor, cursor_to_match)
-
-    while cursor.parent() != cursor_to_match.parent():
-        cursor = cursor.parent()
-        if isinstance(cursor, InvalidCursor):
-            raise CursorNavigationError(
-                "cursor_to_match's parent is not an ancestor of cursor"
-            )
-
-    return cursor
-
-
-def get_stmt_within_scope(cursor, scope):
-    """
-    Gets the statement containing [cursor] that is directly in the provided [scope]
-    """
-    validate_cursors(cursor, scope)
-    assert isinstance(
-        scope, (ForCursor, IfCursor)
-    ), "scope was not an for loop or if statement"
-
-    try:
-        return match_level(cursor, scope.body()[0])
-    except CursorNavigationError:
-        raise CursorNavigationError("scope is not an ancestor of cursor")
-
-
-def get_enclosing_loop(cursor, loop_iter=None):
-    """
-    Gets the enclosing loop with the given [loop_iter]. If [loop_iter] is None,
-    returns the innermost loop enclosing [cursor].
-    """
-    assert not isinstance(cursor, InvalidCursor), "first cursor was an InvalidCursor"
-
-    match_iter = (
-        lambda x: x.name() == loop_iter if loop_iter is not None else lambda x: True
-    )
-
-    while not (isinstance(cursor, ForCursor) and match_iter(cursor)):
-        cursor = cursor.parent()
-        if isinstance(cursor, InvalidCursor):
-            raise CursorNavigationError("no enclosing loop found")
-
-    return cursor
-
-
-def get_top_level_stmt(c):
-    while not isinstance(c.parent(), InvalidCursor):
-        c = c.parent()
-    return c
-
-
-# --------------------------------------------------------------------------- #
-# --------------------------------------------------------------------------- #
 # Internal Functions; Not for Exposure to Users
 
 # helper function to dispatch to constructors
@@ -1079,9 +1000,4 @@ __all__ = [
     "StrideExprCursor",
     #
     "InvalidCursorError",
-    "CursorNavigationError",
-    #
-    "match_level",
-    "get_stmt_within_scope",
-    "get_enclosing_loop",
 ]
