+++ src/exo/stdlib/inspection.py
@@ -5,7 +5,6 @@ from exo.libs.memories import *
 from exo.platforms.x86 import *
 from exo.platforms.neon import *
 from exo.syntax import *
-from exo.stdlib.scheduling import *
 from exo.API_cursors import *
 from exo.stdlib.analysis import *
 
@@ -203,6 +202,43 @@ def is_single_stmt_loop(proc, loop):
     return loop_body_len(proc, loop) == 1
 
 
+# --------------------------------------------------------------------------- #
+# High-level cursor navigation functions
+# --------------------------------------------------------------------------- #
+
+
+class CursorNavigationError(Exception):
+    pass
+
+
+def match_depth(cursor, cursor_to_match):
+    """
+    Lifts [cursor] through the AST until [cursor] and [cursor_to_match] are at
+    the same level, e.g. have the same parent. Returns an InvalidCursorError if
+    [cursor_to_match]'s parent is not an ancestor of [cursor].
+    """
+    assert (
+        cursor.proc() == cursor_to_match.proc()
+    ), "cursors originate from different procs"
+
+    while cursor.parent() != cursor_to_match.parent():
+        cursor = cursor.parent()
+        if isinstance(cursor, InvalidCursor):
+            raise CursorNavigationError(
+                "cursor_to_match's parent is not an ancestor of cursor"
+            )
+
+    return cursor
+
+
+def get_top_level_stmt(proc, c):
+    c = proc.forward(c)
+
+    while not isinstance(c.parent(), InvalidCursor):
+        c = c.parent()
+    return c
+
+
 def get_enclosing_scope(proc, cursor, scope_type):
     cursor = proc.forward(cursor)
     cursor = cursor.parent()
@@ -219,6 +255,20 @@ def get_enclosing_loop(proc, cursor, n=1):
     cursor = proc.forward(cursor)
     for i in range(n):
         cursor = get_enclosing_scope(proc, cursor, ForCursor)
+    return cursor
+
+
+def get_enclosing_loop_by_name(proc, cursor: Cursor, loop_iter: str):
+    """
+    Gets the enclosing loop with the given iteration variable [loop_iter].
+    """
+    cursor = proc.forward(cursor)
+
+    while not (isinstance(cursor, ForCursor) and cursor.name() == loop_iter):
+        cursor = cursor.parent()
+        if isinstance(cursor, InvalidCursor):
+            raise CursorNavigationError("no enclosing loop found")
+
     return cursor
