+++ tests/golden/test_cursors/test_get_enclosing_loop_by_name.txt
